Create planner service #42
Created a planner service and a roslaunch for it. When it is
initialized, the planner will read the contents of mission.json,
and create a new mission.Mission object. It will then solve
the optimization problem to produce an optimal surface on which
plans between waypoints can be generated.

It opens a plot in a separate process of the resulting plan.

Finally, a roslaunch file was made. This is to allow configuration
of the planner through a ros configuration file. This file is in
config/planner_settings.yaml.